Split browser and node tests
